# feat(MM-63936): add mark function in performance metrics manager