<!-- Thank you for contributing a pull request! Here are a few tips to help you:

1. If this is your first contribution, make sure you've read the Contribution Checklist https://developers.mattermost.com/contribute/getting-started/contribution-checklist/
2. Read our blog post about "Submitting Great PRs" https://developers.mattermost.com/blog/2019-01-24-submitting-great-prs
3. Take a look at other repository specific documentation at https://developers.mattermost.com/contribute
-->

#### Summary
<!--
A brief description of what this pull request does.
-->
This PR adds a new mark function to the Performance Metrics Manager to enable more detailed performance tracking with multiple marks. The function allows adding intermediate marks with optional details between the start and end of a metric measurement, providing better insights into the timing of different stages within a performance measurement.

Adding .mark(metricName, detail) at a few key places would create an output as shown below:

```
 LOG  Performance marks for entryRest:
 LOG  Mark 0 => 1:    2674.04ms    [startMetric => after getHasCRTChanged]
 LOG  Mark 1 => 2:     261.07ms    [after getHasCRTChanged => after fetchMyTeams && fetchMe]
 LOG  Mark 2 => 3:       5.57ms    [after fetchMyTeams && fetchMe => before fetchMyChannelsForTeam]
 LOG  Mark 3 => 4:    1061.40ms    [before fetchMyChannelsForTeam => after fetchMyChannelsForTeam]
 LOG  Mark 4 => 5:   20300.07ms    [after fetchMyChannelsForTeam => endMetric]
 LOG  Total:         24302.14ms
 DEBUG  Performance metric: {"metric": "entryRest", "timestamp": 1745946999249, "value": 24302.141329001635}
```

#### Ticket Link
<!--
If this pull request addresses a Help Wanted ticket or fixes a reported issue, please link the relevant GitHub issue, e.g.

  Fixes MM-63936

Otherwise, link the JIRA ticket.
-->

#### Checklist
<!--
Place an '[x]' (no spaces) in all applicable fields. Please remove unrelated fields.
-->
- [x] Added or updated unit tests (required for all new features)
- [ ] Has UI changes
- [ ] Includes text changes and localization file updates
- [ ] Have tested against the 5 core themes to ensure consistency between them.
- [ ] Have run E2E tests by adding label `E2E iOS tests for PR`.

#### Device Information
This PR was tested on: <!-- Device name(s), OS version(s) -->

#### Screenshots
<!--
If the PR includes UI changes, include screenshots/GIFs/Videos (for both iOS and Android if possible).
-->

#### Release Note
<!--
Add a release note for each of the following conditions:

* New features and improvements, including behavioural changes, UI changes
* Bug fixes and fixes of previous known issues
* Deprecation warnings, breaking changes, or compatibility notes

If no release notes are required write NONE. Use past-tense. Newlines are stripped.

Example:

```release-note
Added a new config setting ServiceSettings.FooBar. Added a new column Foo to the Users table.
```

```release-note
NONE
```
-->

```release-note

```